+++ mod.ts
@@ -7,7 +7,7 @@
  * \_____/ \__,_||_| |_||_|/_/\_\
  */
 
-import * as validator from 'modules/validations/mod.ts'
+import * as validatorModule from 'modules/validations/mod.ts'
 
 export * from 'modules/testing/mod.ts'
 export * from 'modules/helpers/mod.ts'
@@ -16,14 +16,12 @@ export * from 'modules/errors/mod.ts'
 export { default as constants } from 'modules/constants/mod.ts'
 export { default as regex } from 'modules/regex/mod.ts'
 
-export {
-  /**
-   * Validator module for BaseRTO-based requests.
-   * This module provides basic validation decorators for string, array, and date types,
-   * using native ECMAScript features compatible with Deno and Microsoft's Reflect Metadata API.
-   * Validations are based on object extensions from BaseRTO.
-   *
-   * @module validator
-   */
-  validator,
-}
+/**
+ * Validator module for BaseRTO-based requests.
+ * This module provides basic validation decorators for string, array, and date types,
+ * using native ECMAScript features compatible with Deno and Microsoft's Reflect Metadata API.
+ * Validations are based on object extensions from BaseRTO.
+ *
+ * @module validator
+ */
+export const validator = validatorModule