+++ src/@tests/unit/utils/strings.test.ts
@@ -1,5 +1,5 @@
+import { capitalize, capitalizeWords, generateHashHex } from 'utils/strings.ts'
 import { assertEquals } from '@std/assert'
-import { capitalize, capitalizeWords } from 'utils/strings.ts'
 
 Deno.test('capitalize should capitalize the first character of a string', () => {
   assertEquals(capitalize('hello'), 'Hello')
@@ -12,4 +12,11 @@ Deno.test('capitalizeWords should capitalize the first letter of each word', () 
   assertEquals(capitalizeWords('the quick brown fox'), 'The Quick Brown Fox')
   assertEquals(capitalizeWords('hello'), 'Hello')
   assertEquals(capitalizeWords(''), '') // Empty string should return empty string
+})
+
+Deno.test('generate hash', async () => {
+  assertEquals(
+    await generateHashHex('text value'),
+    'ffabc3fe711165f0a9acd034603b7c4dace74cbe643409a89ab62c9c3a87c4a8',
+  )
 })