+++ src/modules/validations/decorators/strings/length.ts
@@ -3,7 +3,7 @@ import type { ValidationDecorator, ValidationDecoratorDefinition } from 'typings
 import { defineValidationDecorator } from 'modules/validations/base/definitions/decorators.ts'
 
 /**
- * has valid length
+ * Validate if has valid length
  *
  * @param value
  * @param min